Convert Bug8579Test to standard rule test and assertType test
Move the Bug8579 regression test from a custom Analyser test to a
standard RuleTestCase for InstantiationRule with
rememberPossiblyImpureFunctionValues: false config.

Convert the class_exists false-not-remembered test to assertType
assertions in the existing DoNotRememberPossiblyImpureFunctionValues
test data file.
