+++ tests/PHPStan/Rules/Classes/data/bug-8579.php
