+++ tests/PHPStan/Rules/Classes/doNotRememberPossiblyImpureValues.neon
@@ -1,0 +1,2 @@
+parameters:
+	rememberPossiblyImpureFunctionValues: false
